bots/requirements.txt
python-telegram-bot>=20.0
lark-oapi>=1.4
flask>=3.0
requests>=2.28
